+++ packages/common/src/visualizations/types/index.ts
@@ -67,6 +67,7 @@ export type PivotChartData = {
     results: RawResultRow[];
     indexColumn: { reference: string; type: string } | undefined;
     valuesColumns: string[];
+    columns: VizSqlColumn[];
 };
 
 export type VizCartesianChartOptions = {
@@ -80,22 +81,24 @@ export type VizPieChartOptions = {
     metricFieldOptions: VizValuesLayoutOptions[];
 };
 
+export type VizColumnConfig = {
+    visible: boolean;
+    reference: string;
+    label: string;
+    frozen: boolean;
+    order?: number;
+    aggregation?: VizAggregationOptions;
+};
+
+export type VizColumnsConfig = Record<string, VizColumnConfig>;
+
+export type VizTableColumnsConfig = {
+    columns: VizColumnsConfig;
+};
+
 // TODO: FIXME!! it should be a common type!
 export type VizTableOptions = {
     defaultColumnConfig: VizTableColumnsConfig['columns'] | undefined;
-};
-
-export type VizTableColumnsConfig = {
-    columns: {
-        [key: string]: {
-            visible: boolean;
-            reference: string;
-            label: string;
-            frozen: boolean;
-            order?: number;
-            aggregation?: VizAggregationOptions;
-        };
-    };
 };
 
 export type VizBaseConfig = {