+++ packages/frontend/src/features/sqlRunner/runners/SqlRunnerResultsRunner.ts
@@ -31,7 +31,7 @@ type PivotQueryFn = (
     args: SqlRunnerPivotQueryBody & {
         projectUuid: string;
     },
-) => Promise<PivotChartData>;
+) => Promise<Omit<PivotChartData, 'columns'>>;
 
 const pivotQueryFn: PivotQueryFn = async ({ projectUuid, ...args }) => {
     const scheduledJob = await schedulePivotSqlJob({
@@ -78,8 +78,10 @@ export class SqlRunnerResultsRunner extends ResultsRunner {
                 results: [],
                 indexColumn: undefined,
                 valuesColumns: [],
+                columns: [],
             };
         }
+
         const pivotResults = await pivotQueryFn({
             projectUuid,
             slug,
@@ -97,10 +99,20 @@ export class SqlRunnerResultsRunner extends ResultsRunner {
             limit,
         });
 
+        const columns: VizSqlColumn[] = [
+            ...(pivotResults.indexColumn?.reference
+                ? [pivotResults.indexColumn.reference]
+                : []),
+            ...pivotResults.valuesColumns,
+        ].map((field) => ({
+            reference: field,
+        }));
+
         return {
             results: pivotResults.results,
             indexColumn: pivotResults.indexColumn,
             valuesColumns: pivotResults.valuesColumns,
+            columns,
         };
     }
 }