+++ packages/warehouses/src/warehouseClients/BigqueryWarehouseClient.ts
@@ -114,7 +114,7 @@ export class BigqueryWarehouseClient extends WarehouseBaseClient<CreateBigqueryC
         super(credentials);
         try {
             this.client = new BigQuery({
-                projectId: credentials.project,
+                projectId: credentials.executionProject || credentials.project,
                 location: credentials.location,
                 maxRetries: credentials.retries,
                 credentials: credentials.keyfileContents,
@@ -223,19 +223,12 @@ export class BigqueryWarehouseClient extends WarehouseBaseClient<CreateBigqueryC
             table: string;
         }[],
     ) {
-        const databaseClients: { [client: string]: BigQuery } = {};
         const tablesMetadataPromises: Promise<
             [string, string, string, TableSchema] | undefined
         >[] = requests.map(({ database, schema, table }) => {
-            databaseClients[database] =
-                databaseClients[database] ||
-                new BigQuery({
-                    projectId: database,
-                    location: this.credentials.location,
-                    maxRetries: this.credentials.retries,
-                    credentials: this.credentials.keyfileContents,
-                });
-            const dataset = databaseClients[database].dataset(schema);
+            const dataset: Dataset = new Dataset(this.client, schema, {
+                projectId: database,
+            });
             return BigqueryWarehouseClient.getTableMetadata(
                 dataset,
                 table,
@@ -297,13 +290,7 @@ export class BigqueryWarehouseClient extends WarehouseBaseClient<CreateBigqueryC
     }
 
     async getAllTables() {
-        const client = new BigQuery({
-            projectId: this.credentials.project,
-            location: this.credentials.location,
-            maxRetries: this.credentials.retries,
-            credentials: this.credentials.keyfileContents,
-        });
-        const [datasets] = await client.getDatasets();
+        const [datasets] = await this.client.getDatasets();
         const datasetTablesResponses = await Promise.all(
             datasets.map((d) => d.getTables()),
         );
@@ -321,13 +308,9 @@ export class BigqueryWarehouseClient extends WarehouseBaseClient<CreateBigqueryC
         schema: string,
         database?: string,
     ): Promise<WarehouseCatalog> {
-        const client = new BigQuery({
-            projectId: database || this.credentials.project,
-            location: this.credentials.location,
-            maxRetries: this.credentials.retries,
-            credentials: this.credentials.keyfileContents,
+        const dataset: Dataset = new Dataset(this.client, schema, {
+            projectId: database,
         });
-        const dataset = client.dataset(schema);
         const schemas = await BigqueryWarehouseClient.getTableMetadata(
             dataset,
             tableName,